+++ crates/block-producer/src/block_builder/mod.rs
@@ -170,7 +170,7 @@ impl BlockBuilder {
                 .input_notes()
                 .iter()
                 .cloned()
-                .filter_map(|note| note.header().map(NoteHeader::id))
+                .filter_map(|note| note.header().map(NoteHeader::commitment))
         });
         let block_references_iter =
             batch_iter.clone().map(Deref::deref).map(ProvenBatch::reference_block_num);