+++ crates/block-producer/src/store/mod.rs
@@ -11,7 +11,7 @@ use miden_node_utils::formatting::format_opt;
 use miden_objects::Word;
 use miden_objects::account::AccountId;
 use miden_objects::block::{BlockHeader, BlockInputs, BlockNumber, ProvenBlock};
-use miden_objects::note::{NoteId, Nullifier};
+use miden_objects::note::Nullifier;
 use miden_objects::transaction::ProvenTransaction;
 use miden_objects::utils::Serializable;
 use tracing::{debug, info, instrument};
@@ -34,10 +34,10 @@ pub struct TransactionInputs {
     ///
     /// We use `NonZeroU32` as the wire format uses 0 to encode none.
     pub nullifiers: HashMap<Nullifier, Option<NonZeroU32>>,
-    /// Unauthenticated notes which are present in the store.
+    /// Unauthenticated note commitments which are present in the store.
     ///
     /// These are notes which were committed _after_ the transaction was created.
-    pub found_unauthenticated_notes: HashSet<NoteId>,
+    pub found_unauthenticated_notes: HashSet<Word>,
     /// The current block height.
     pub current_block_height: BlockNumber,
 }
@@ -97,7 +97,7 @@ impl TryFrom<proto::block_producer_store::TransactionInputs> for TransactionInpu
         let found_unauthenticated_notes = response
             .found_unauthenticated_notes
             .into_iter()
-            .map(|digest| Ok(Word::try_from(digest)?.into()))
+            .map(Word::try_from)
             .collect::<Result<_, ConversionError>>()?;
 
         let current_block_height = response.block_height.into();
@@ -167,7 +167,7 @@ impl StoreClient {
             nullifiers: proven_tx.nullifiers().map(Into::into).collect(),
             unauthenticated_notes: proven_tx
                 .unauthenticated_notes()
-                .map(|note| note.id().into())
+                .map(|note| note.commitment().into())
                 .collect(),
         };
 
@@ -207,7 +207,7 @@ impl StoreClient {
         &self,
         updated_accounts: impl Iterator<Item = AccountId> + Send,
         created_nullifiers: impl Iterator<Item = Nullifier> + Send,
-        unauthenticated_notes: impl Iterator<Item = NoteId> + Send,
+        unauthenticated_notes: impl Iterator<Item = Word> + Send,
         reference_blocks: impl Iterator<Item = BlockNumber> + Send,
     ) -> Result<BlockInputs, StoreError> {
         let request = tonic::Request::new(proto::block_producer_store::BlockInputsRequest {
@@ -228,11 +228,11 @@ impl StoreClient {
     pub async fn get_batch_inputs(
         &self,
         block_references: impl Iterator<Item = (BlockNumber, Word)> + Send,
-        notes: impl Iterator<Item = NoteId> + Send,
+        note_commitments: impl Iterator<Item = Word> + Send,
     ) -> Result<BatchInputs, StoreError> {
         let request = tonic::Request::new(proto::block_producer_store::BatchInputsRequest {
             reference_blocks: block_references.map(|(block_num, _)| block_num.as_u32()).collect(),
-            note_ids: notes.map(proto::primitives::Digest::from).collect(),
+            note_commitments: note_commitments.map(proto::primitives::Digest::from).collect(),
         });
 
         let store_response = self.client.clone().get_batch_inputs(request).await?.into_inner();