+++ crates/store/src/db/schema.rs
@@ -61,6 +61,7 @@ diesel::table! {
         batch_index -> Integer,
         note_index -> Integer,
         note_id -> Binary,
+        note_commitment -> Binary,
         note_type -> Integer,
         sender -> Binary,
         tag -> Integer,